swarm/tools/quantum_escalation_tool.py
"""
ADK Tool: Quantum Escalation (Real ADK API)
Agent uses this to trigger quantum analysis
"""

from google.adk.tools import BaseTool
from typing import Any
from pydantic import BaseModel, Field

class QuantumEscalationInput(BaseModel):
    """Input schema for quantum escalation"""
    consensus_data: dict = Field(description="Consensus results from swarm")
    swarm_evidence: dict = Field(description="Collective evidence from all agents")

class EscalateToQuantumTool(BaseTool):
    """
    Tool for escalating to quantum analysis layer
    Used when swarm reaches consensus
    """
    
    input_schema: type[BaseModel] = QuantumEscalationInput
    
    def __init__(self, quantum_service=None, **kwargs):
        super().__init__(
            name="escalate_to_quantum",
            description="""Trigger quantum analysis after swarm consensus.
    
Use this tool when:
- Consensus reached among agents
- Pattern requires deep analysis
- Resource optimization needed

This invokes TensorFlow Quantum circuits for:
- Pattern detection
- Resource allocation
- Causality analysis
""",
            **kwargs
        )
        self.quantum_service = quantum_service
    
    async def run(self, consensus_data: dict, swarm_evidence: dict, **kwargs) -> dict[str, Any]:
        """
        Trigger quantum analysis
        """
        if not self.quantum_service:
            return {"error": "Quantum service not configured"}
        
        try:
            # Call quantum service
            quantum_result = await self.quantum_service.analyze_outbreak_pattern(
                swarm_data=swarm_evidence
            )
            
            # Also optimize resources if outbreak likely
            if quantum_result.get('outbreak_probability', 0) > 0.7:
                villages = swarm_evidence.get('villages', [])
                allocation = await self.quantum_service.optimize_resource_allocation(
                    villages=villages,
                    resources={'ors': 1000, 'staff': 50, 'kits': 500}
                )
                quantum_result['resource_allocation'] = allocation
            
            return {
                "quantum_analysis_complete": True,
                "triggered_by_consensus": True,
                "results": quantum_result
            }
        except Exception as e:
            return {
                "error": str(e),
                "quantum_analysis_complete": False
            }
