requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0

# Google AI & ADK
google-generativeai==0.3.1
google-cloud-aiplatform==1.38.0
google-adk==1.21.0                    # ✅ AI Development Kit (REAL PACKAGE!)

# Quantum Computing
tensorflow==2.16.2
tensorflow-quantum
cirq==1.3.0
cirq-core==1.3.0

# Database & Cache
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
redis==5.0.1
asyncpg==0.29.0

# Data Processing
pandas==2.1.4
numpy==1.24.3
scipy==1.11.4
scikit-learn==1.3.2

# Utilities
pyyaml==6.0.1
aiofiles==23.2.1
python-dateutil==2.8.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
